+++ app/Http/Controllers/SuperAdmin/HrNotificationController.php
@@ -10,8 +10,18 @@ class HrNotificationController extends Controller
 {
     public function index()
     {
-        $notification = HrNotification::get();
-        return view('superadmin.hr_notification.index',compact('notification'));
+        $userId = auth()->id();
+        $buildingAdmin = auth()->guard('building_admin')->user();
+
+        $notification = collect();
+        if ($buildingAdmin) {
+            $notification = HrNotification::where('building_admin_id', $buildingAdmin->id)->get();
+        } else {
+            $notification = HrNotification::where('user_id', $userId)->get();
+        }
+
+//        $notification = HrNotification::where('user_id',auth()->id())->get();
+        return view('superadmin.hr_notification.index', compact('notification'));
     }
 
     public function create()
@@ -27,11 +37,24 @@ class HrNotificationController extends Controller
             'date' => 'required',
             'status' => 'required',
         ]);
+
+
+        $userId = null;
+        if (auth()->user()) {
+            $userId = auth()->user()->id;
+        } elseif (auth()->guard('building_admin')->check()) {
+
+            $buildingAdmin = auth()->guard('building_admin')->user();
+            $userId = $buildingAdmin->user_id;
+        }
+
         HrNotification::create([
-           'notification_type' => $request->notification_type,
-           'message' => $request->message,
-           'date' => $request->date,
-           'status' => $request->status,
+            'user_id' => $userId,
+            'building_admin_id' => auth()->guard('building_admin')->id(),
+            'notification_type' => $request->notification_type,
+            'message' => $request->message,
+            'date' => $request->date,
+            'status' => $request->status,
         ]);
         return redirect()->route('hr_notification.index');
 
@@ -39,30 +62,30 @@ class HrNotificationController extends Controller
 
     public function edit($id)
     {
-        $notifications = HrNotification::findOrFail($id);
-        return view('superadmin.hr_notification.edit',compact('notifications'));
+        $notifications = HrNotification::where('user_id',auth()->id())->findOrFail($id);
+        return view('superadmin.hr_notification.edit', compact('notifications'));
     }
 
-    public function update(Request $request,$id)
+    public function update(Request $request, $id)
     {
-         $request->validate([
+        $request->validate([
             'notification_type' => 'required',
             'message' => 'required',
             'date' => 'required',
             'status' => 'required',
         ]);
         HrNotification::findOrFail($id)->update([
-           'notification_type' => $request->notification_type,
-           'message' => $request->message,
-           'date' => $request->date,
-           'status' => $request->status,
+            'notification_type' => $request->notification_type,
+            'message' => $request->message,
+            'date' => $request->date,
+            'status' => $request->status,
         ]);
         return redirect()->route('hr_notification.index');
     }
 
     public function destroy($id)
     {
-        HrNotification::findOrFail($id)->delete();
+        HrNotification::where('user_id',auth()->id())->findOrFail($id)->delete();
         return redirect()->back();
     }
