+++ app/Http/Controllers/SuperAdmin/ParkingController.php
@@ -15,12 +15,32 @@ class ParkingController extends Controller
 
     public function index()
     {
-        $parking = Parking::with('allotment')->get();
+        $userId = auth()->id();
+        $buildingAdmin = auth()->guard('building_admin')->user();
+
+        $parking = collect();
+        if ($buildingAdmin) {
+            $parking = Parking::where('building_admin_id', $buildingAdmin->id)->with(['allotment'])->get();
+        } else {
+            $parking = Parking::where('user_id', $userId)->with(['allotment'])->get();
+        }
+
+//        $parking = Parking::where('user_id',auth()->id())->with('allotment')->get();
         return view('superadmin.parking.index',compact('parking'));
     }
     public function create()
     {
-        $allotments = Allotment::get();
+        $userId = auth()->id();
+        $buildingAdmin = auth()->guard('building_admin')->user();
+
+        $allotments = collect();
+        if ($buildingAdmin) {
+            $allotments = Allotment::where('building_admin_id', $buildingAdmin->id)->get();
+        } else {
+            $allotments = Allotment::where('user_id', $userId)->get();
+        }
+
+//        $allotments = Allotment::where('user_id',auth()->id())->get();
         return view('superadmin.parking.create',compact('allotments'));
     }
 
@@ -32,7 +52,18 @@ class ParkingController extends Controller
            'vehicle_number' => 'required',
            'parking_status' => 'required',
         ]);
+        $userId = null;
+        if (auth()->user()) {
+            $userId = auth()->user()->id;
+        } elseif (auth()->guard('building_admin')->check()) {
+
+            $buildingAdmin = auth()->guard('building_admin')->user();
+            $userId = $buildingAdmin->user_id;
+        }
+
         Parking::create([
+            'user_id' => $userId,
+            'building_admin_id' => auth()->guard('building_admin')->id(),
            'allotment_id' => $request->allotment_id,
            'parking_space_number' => $request->parking_space_number,
            'vehicle_number' => $request->vehicle_number,
@@ -43,8 +74,8 @@ class ParkingController extends Controller
     }
     public function edit($id)
     {
-        $parking = Parking::findOrFail($id);
-        $allotments = Allotment::get();
+        $parking = Parking::where('user_id',auth()->id())->findOrFail($id);
+        $allotments = Allotment::where('user_id',auth()->id())->get();
         return view('superadmin.parking.edit',compact('parking','allotments'));
     }
     public function update(Request $request,$id)
@@ -65,7 +96,7 @@ class ParkingController extends Controller
     }
     public function destroy($id)
     {
-        Parking::findOrFail($id)->delete();
+        Parking::where('user_id',auth()->id())->findOrFail($id)->delete();
         return redirect()->back();
     }
 }
