+++ app/Http/Controllers/SuperAdmin/PayrollController.php
@@ -13,33 +13,63 @@ class PayrollController extends Controller
 
     public function index()
     {
-        $payrolls = Payroll::with('employee')->get();
+        $userId = auth()->id();
+        $buildingAdmin = auth()->guard('building_admin')->user();
+
+        $payrolls = collect();
+        if ($buildingAdmin) {
+            $payrolls = Payroll::where('building_admin_id', $buildingAdmin->id)->with(['employee'])->get();
+        } else {
+            $payrolls = Payroll::where('user_id', $userId)->with(['employee'])->get();
+        }
+
+//        $payrolls = Payroll::where('user_id',auth()->id())->with('employee')->get();
         return view('superadmin.payroll.index', compact('payrolls'));
     }
 
     public function create()
     {
-        $employees = Employees::get();
-        return view('superadmin.payroll.create',compact('employees'));
+        $userId = auth()->id();
+        $buildingAdmin = auth()->guard('building_admin')->user();
+
+        $employees = collect();
+        if ($buildingAdmin) {
+            $employees = Employees::where('building_admin_id', $buildingAdmin->id)->get();
+        } else {
+            $employees = Employees::where('user_id', $userId)->get();
+        }
+
+//        $employees = Employees::where('user_id',auth()->id())->get();
+        return view('superadmin.payroll.create', compact('employees'));
     }
 
     public function store(Request $request)
     {
         $request->validate([
-           'employee_id' => 'required',
-           'salary_amount' => 'required',
-           'deducation'  => 'required',
-           'net_salary' => 'required',
-           'pay_date' => 'required|Date'
+            'employee_id' => 'required',
+            'salary_amount' => 'required',
+            'deducation' => 'required',
+            'net_salary' => 'required',
+            'pay_date' => 'required|Date'
         ]);
+        $userId = null;
+        if (auth()->user()) {
+            $userId = auth()->user()->id;
+        } elseif (auth()->guard('building_admin')->check()) {
+
+            $buildingAdmin = auth()->guard('building_admin')->user();
+            $userId = $buildingAdmin->user_id;
+        }
 
         Payroll::create([
-           'employee_id' => $request->employee_id,
-           'salary_amount' => $request->salary_amount,
-           'deducation' => $request->deducation,
-           'net_salary' => $request->net_salary,
-           'pay_date' => $request->pay_date,
-           'created_at' => Carbon::now(),
+            'user_id' => $userId,
+            'building_admin_id' => auth()->guard('building_admin')->id(),
+            'employee_id' => $request->employee_id,
+            'salary_amount' => $request->salary_amount,
+            'deducation' => $request->deducation,
+            'net_salary' => $request->net_salary,
+            'pay_date' => $request->pay_date,
+            'created_at' => Carbon::now(),
         ]);
 
         return redirect()->route('payroll.index');
@@ -47,29 +77,29 @@ class PayrollController extends Controller
 
     public function edit($id)
     {
-        $payroll = Payroll::findOrFail($id);
+        $payroll = Payroll::where('user_id', auth()->id())->findOrFail($id);
         $employees = Employees::get();
-        return view('superadmin.payroll.edit',compact('payroll','employees'));
+        return view('superadmin.payroll.edit', compact('payroll', 'employees'));
     }
 
-    public function update(Request $request ,$id)
+    public function update(Request $request, $id)
     {
         $payroll_id = $request->id;
 
         $request->validate([
-           'employee_id' => 'required',
-           'salary_amount' => 'required',
-           'deducation'  => 'required',
-           'net_salary' => 'required',
-           'pay_date' => 'required|Date'
+            'employee_id' => 'required',
+            'salary_amount' => 'required',
+            'deducation' => 'required',
+            'net_salary' => 'required',
+            'pay_date' => 'required|Date'
         ]);
 
         Payroll::findOrFail($payroll_id)->update([
-           'employee_id' => $request->employee_id,
-           'salary_amount' => $request->salary_amount,
-           'deducation' => $request->deducation,
-           'net_salary' => $request->net_salary,
-           'pay_date' => $request->pay_date,
+            'employee_id' => $request->employee_id,
+            'salary_amount' => $request->salary_amount,
+            'deducation' => $request->deducation,
+            'net_salary' => $request->net_salary,
+            'pay_date' => $request->pay_date,
             'update_at' => Carbon::now(),
         ]);
 
@@ -78,7 +108,7 @@ class PayrollController extends Controller
 
     public function destroy($id)
     {
-        Payroll::findOrFail($id)->delete();
+        Payroll::where('user_id', auth()->id())->findOrFail($id)->delete();
         return redirect()->back();
     }
 }
